refactor(capsule): move component-ext to capsule/component-capsule
