+++ src/consumer/component/consumer-component.ts
@@ -71,7 +71,7 @@ import Isolator, { IsolateOptions } from '../../environment/isolator';
 import { stripSharedDirFromPath } from '../component-ops/manipulate-dir';
 import ComponentsPendingImport from '../component-ops/exceptions/components-pending-import';
 import ExtensionIsolateResult from '../../legacy-extensions/extension-isolate-result';
-import { ComponentCapsule } from '../../extensions/capsule-ext';
+import { ComponentCapsule } from '../../extensions/capsule/component-capsule';
 
 export type customResolvedPath = { destinationPath: PathLinux; importSource: string };
