+++ src/consumer/component/sources/data-to-persist.ts
@@ -1,7 +1,7 @@
 import * as path from 'path';
 import Bluebird from 'bluebird';
 import fs from 'fs-extra';
-import { ComponentCapsule } from '../../../extensions/capsule-ext';
+import { ComponentCapsule } from '../../../extensions/capsule/component-capsule';
 import AbstractVinyl from './abstract-vinyl';
 import Symlink from '../../../links/symlink';
 import logger from '../../../logger/logger';
