+++ src/environment/isolator.ts
@@ -3,7 +3,7 @@ import * as path from 'path';
 import semver from 'semver';
 import pMapSeries from 'p-map-series';
 import { runModule } from 'librarian';
-import { ComponentCapsule } from '../extensions/capsule-ext';
+import { ComponentCapsule } from '../extensions/capsule/component-capsule';
 import createCapsule from './capsule-factory';
 import Consumer from '../consumer/consumer';
 import { Scope, ComponentWithDependencies } from '../scope';
