+++ src/extensions/capsule/component-capsule/component-capsule.ts
@@ -4,8 +4,8 @@ import { NodeFS } from '@teambit/any-fs';
 import _ from 'lodash';
 import librarian from 'librarian';
 import FsContainer, { BitExecOption } from './container';
-import BitId from '../../bit-id/bit-id';
-import BitContainerFactory from '../network/orchestrator/bit-container-factory';
+import BitId from '../../../bit-id/bit-id';
+import BitContainerFactory from '../../network/orchestrator/bit-container-factory';
 
 export default class ComponentCapsule extends Capsule<Exec, NodeFS> {
   private _wrkDir: string;
