+++ src/extensions/capsule/component-capsule/container.ts
