+++ src/extensions/capsule/component-capsule/index.ts
