+++ src/extensions/compile/compile.ts
@@ -4,7 +4,7 @@ import { BitId } from '../../bit-id';
 import { ResolvedComponent } from '../workspace/resolved-component';
 import buildComponent from '../../consumer/component-ops/build-component';
 import { Component } from '../component';
-import { ComponentCapsule } from '../capsule-ext';
+import { ComponentCapsule } from '../capsule/component-capsule';
 import DataToPersist from '../../consumer/component/sources/data-to-persist';
 import { Scripts } from '../scripts';
 import { IdsAndScripts } from '../scripts/ids-and-scripts';
