+++ src/extensions/network/capsule-list.ts
@@ -1,4 +1,4 @@
 import BitIdAndValueArray from '../../bit-id/bit-id-and-value-array';
-import { ComponentCapsule } from '../capsule-ext';
+import { ComponentCapsule } from '../capsule/component-capsule';
 
 export default class CapsuleList extends BitIdAndValueArray<ComponentCapsule> {}
