+++ src/extensions/network/orchestrator/bit-container-factory.ts
@@ -1,6 +1,6 @@
 import { ContainerFactory, Container, Exec } from '@teambit/capsule';
 import { AnyFS } from '@teambit/any-fs';
-import { FsContainer, BitContainerConfig } from '../../capsule-ext';
+import { FsContainer, BitContainerConfig } from '../../capsule/component-capsule';
 
 export default class BitContainerFactory implements ContainerFactory<Exec, AnyFS> {
   createContainer(options?: BitContainerConfig): Promise<Container<Exec, AnyFS>> {
