+++ src/extensions/workspace/resolved-component.ts
@@ -1,5 +1,5 @@
 import { Component } from '../component';
-import { ComponentCapsule } from '../capsule-ext';
+import { ComponentCapsule } from '../capsule/component-capsule';
 
 export class ResolvedComponent {
   constructor(readonly component: Component, readonly capsule: ComponentCapsule) {}
